# server/swagger.yaml
...
openapi: 3.0.0
info:
  title: softcloud
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /volume/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Volume
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Volume
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volume"
    post:
      summary: CRUD ACTION update
      tags:
        - Volume
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volume"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Volume"
  /volume/findByenvironments/{key}:
    get:
      summary: CRUD ACTION findByenvironments
      tags:
        - Volume
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Volume"
    
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Volume:
      properties:
        _id:
          type: string
        InternalConfig:
          type: string
        MountPoints:
          type: string
        Name:
          type: string
        Size:
          type: number
      required:
        -  _id